Rename e14_slash_commands to e06_sample_bot_structure
