--- examples/e06_sample_bot_structure/.env.example
@@ -1,10 +1,0 @@
-# This declares an environment variable named "DISCORD_TOKEN" with the given
-# value. When calling `dotenv::dotenv()`, it will read the `.env` file and parse
-# these key-value pairs and insert them into the environment.
-#
-# Environment variables are separated by newlines and must not have space
-# around the equals sign (`=`).
-DISCORD_TOKEN=put your token here
-# Declares the level of logging to use. Read the documentation for the `log`
-# and `env_logger` crates for more information.
-RUST_LOG=debug
