+++ examples/e06_sample_bot_structure/README.md
