--- examples/e06_sample_bot_structure/src/commands/math.rs
@@ -1,16 +1,0 @@
-use serenity::framework::standard::macros::command;
-use serenity::framework::standard::{Args, CommandResult};
-use serenity::model::prelude::*;
-use serenity::prelude::*;
-
-#[command]
-pub async fn multiply(ctx: &Context, msg: &Message, mut args: Args) -> CommandResult {
-    let one = args.single::<f64>()?;
-    let two = args.single::<f64>()?;
-
-    let product = one * two;
-
-    msg.channel_id.say(&ctx.http, product.to_string()).await?;
-
-    Ok(())
-}
