+++ examples/e06_sample_bot_structure/src/commands/ping.rs
