+++ examples/e06_sample_bot_structure/src/commands/wonderful_command.rs
