+++ examples/e06_sample_bot_structure/src/main.rs
@@ -1,114 +1,90 @@
-//! Requires the 'framework' feature flag be enabled in your project's `Cargo.toml`.
-//!
-//! This can be enabled by specifying the feature in the dependency section:
-//!
-//! ```toml
-//! [dependencies.serenity]
-//! git = "https://github.com/serenity-rs/serenity.git"
-//! features = ["framework", "standard_framework"]
-//! ```
-#![allow(deprecated)] // We recommend migrating to poise, instead of using the standard command framework.
 mod commands;
 
-use std::collections::HashSet;
 use std::env;
-use std::sync::{Arc, OnceLock};
 
 use serenity::async_trait;
-use serenity::framework::standard::macros::group;
-use serenity::framework::standard::Configuration;
-use serenity::framework::StandardFramework;
-use serenity::gateway::ShardManager;
-use serenity::http::Http;
-use serenity::model::event::ResumedEvent;
+use serenity::builder::{CreateInteractionResponse, CreateInteractionResponseMessage};
+use serenity::model::application::{Command, Interaction};
 use serenity::model::gateway::Ready;
+use serenity::model::id::GuildId;
 use serenity::prelude::*;
-use tracing::{error, info};
-
-use crate::commands::math::*;
-use crate::commands::meta::*;
-use crate::commands::owner::*;
-
-struct UserData {
-    shard_manager: OnceLock<Arc<ShardManager>>,
-}
 
 struct Handler;
 
 #[async_trait]
 impl EventHandler for Handler {
-    async fn ready(&self, _: &Context, ready: &Ready) {
-        info!("Connected as {}", ready.user.name);
+    async fn interaction_create(&self, ctx: &Context, interaction: &Interaction) {
+        if let Interaction::Command(command) = interaction {
+            println!("Received command interaction: {command:#?}");
+
+            let content = match command.data.name.as_str() {
+                "ping" => Some(commands::ping::run(&command.data.options())),
+                "id" => Some(commands::id::run(&command.data.options())),
+                "attachmentinput" => Some(commands::attachmentinput::run(&command.data.options())),
+                "modal" => {
+                    commands::modal::run(ctx, command).await.unwrap();
+                    None
+                },
+                _ => Some("not implemented :(".to_string()),
+            };
+
+            if let Some(content) = content {
+                let data = CreateInteractionResponseMessage::new().content(content);
+                let builder = CreateInteractionResponse::Message(data);
+                if let Err(why) = command.create_response(&ctx.http, builder).await {
+                    println!("Cannot respond to slash command: {why}");
+                }
+            }
+        }
     }
 
-    async fn resume(&self, _: &Context, _: &ResumedEvent) {
-        info!("Resumed");
+    async fn ready(&self, ctx: &Context, ready: &Ready) {
+        println!("{} is connected!", ready.user.name);
+
+        let guild_id = GuildId::new(
+            env::var("GUILD_ID")
+                .expect("Expected GUILD_ID in environment")
+                .parse()
+                .expect("GUILD_ID must be an integer"),
+        );
+
+        let commands = guild_id
+            .set_commands(&ctx.http, &[
+                commands::ping::register(),
+                commands::id::register(),
+                commands::welcome::register(),
+                commands::numberinput::register(),
+                commands::attachmentinput::register(),
+                commands::modal::register(),
+            ])
+            .await;
+
+        println!("I now have the following guild slash commands: {commands:#?}");
+
+        let guild_command =
+            Command::create_global_command(&ctx.http, commands::wonderful_command::register())
+                .await;
+
+        println!("I created the following global slash command: {guild_command:#?}");
     }
 }
 
-#[group]
-#[commands(multiply, ping, quit)]
-struct General;
-
 #[tokio::main]
 async fn main() {
-    // This will load the environment variables located at `./.env`, relative to the CWD.
-    // See `./.env.example` for an example on how to structure this.
-    dotenv::dotenv().expect("Failed to load .env file");
-
-    // Initialize the logger to use environment variables.
-    //
-    // In this case, a good default is setting the environment variable `RUST_LOG` to `debug`.
-    tracing_subscriber::fmt::init();
-
+    // Configure the client with your Discord bot token in the environment.
     let token = env::var("DISCORD_TOKEN").expect("Expected a token in the environment");
 
-    let http = Http::new(&token);
-
-    // We will fetch your bot's owners and id
-    let (owners, _bot_id) = match http.get_current_application_info().await {
-        Ok(info) => {
-            let mut owners = HashSet::new();
-            if let Some(owner) = &info.owner {
-                owners.insert(owner.id);
-            }
-
-            (owners, info.id)
-        },
-        Err(why) => panic!("Could not access application info: {:?}", why),
-    };
-
-    // Create the framework
-    let framework = StandardFramework::new().group(&GENERAL_GROUP);
-    framework.configure(Configuration::new().owners(owners).prefix("~"));
-
-    let user_data = UserData {
-        shard_manager: OnceLock::new(),
-    };
-
-    let intents = GatewayIntents::GUILD_MESSAGES
-        | GatewayIntents::DIRECT_MESSAGES
-        | GatewayIntents::MESSAGE_CONTENT;
-    let mut client = Client::builder(&token, intents)
-        .framework(framework)
+    // Build our client.
+    let mut client = Client::builder(&token, GatewayIntents::empty())
         .event_handler(Handler)
-        .data(Arc::new(user_data) as _)
         .await
-        .expect("Err creating client");
+        .expect("Error creating client");
 
-    {
-        let data = client.data::<UserData>();
-        let _ = data.shard_manager.set(client.shard_manager.clone());
-    }
-
-    let shard_manager = client.shard_manager.clone();
-
-    tokio::spawn(async move {
-        tokio::signal::ctrl_c().await.expect("Could not register ctrl+c handler");
-        shard_manager.shutdown_all().await;
-    });
-
+    // Finally, start a single shard, and start listening to events.
+    //
+    // Shards will automatically attempt to reconnect, and will perform exponential backoff until
+    // it reconnects.
     if let Err(why) = client.start().await {
-        error!("Client error: {:?}", why);
+        println!("Client error: {why:?}");
     }
 }
