--- examples/e14_slash_commands/src/commands/mod.rs
@@ -1,7 +1,0 @@
-pub mod attachmentinput;
-pub mod id;
-pub mod modal;
-pub mod numberinput;
-pub mod ping;
-pub mod welcome;
-pub mod wonderful_command;
